add controller calls, cluster statistic and fix pom.xml

+++ src/main/java/com/percolation/domain/statistic/ClusterStatistic.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ClusterStatistic {
-    private int averageAmountClusters;
+    private double averageAmountClusters;
     private double averageSizeCluster;
     private double averageDistanceBetweenClusters;
 }
